LWOLoader: dev branch factorization
